Warning fixes

+++ src/game/components/well/Render.cpp
@@ -19,14 +19,14 @@ Render::Render()
 void Render::drawBackground(const Well&, GraphicsContext&, int draw_offset_x, int draw_offset_y) const
 {
     const auto& cell = MinoStorage::getMatrixCell();
-    for (size_t col = 0; col < 10; col++) {
+    for (int col = 0; col < 10; col++) {
         cell->drawPartial(top_row_cliprect, {
             draw_offset_x + col * Mino::texture_size_px, draw_offset_y,
             Mino::texture_size_px, top_row_height});
     }
     draw_offset_y += top_row_height;
-    for (size_t row = 0; row < 20; row++) {
-        for (size_t col = 0; col < 10; col++) {
+    for (unsigned row = 0; row < 20; row++) {
+        for (unsigned col = 0; col < 10; col++) {
             cell->draw(draw_offset_x + col * Mino::texture_size_px,
                        draw_offset_y + row * Mino::texture_size_px);
         }
@@ -36,7 +36,7 @@ void Render::drawBackground(const Well&, GraphicsContext&, int draw_offset_x, in
 void Render::drawContent(const Well& well, GraphicsContext& gcx, int draw_offset_x, int draw_offset_y) const
 {
     // Draw board Minos
-    for (size_t col = 0; col < 10; col++) {
+    for (int col = 0; col < 10; col++) {
         const auto& cell = well.matrix.at(1).at(col);
         if (cell) {
             cell->drawPartial(top_row_cliprect, {
@@ -45,8 +45,8 @@ void Render::drawContent(const Well& well, GraphicsContext& gcx, int draw_offset
         }
     }
     draw_offset_y += top_row_height;
-    for (size_t row = 0; row < 20; row++) {
-        for (size_t col = 0; col < 10; col++) {
+    for (unsigned row = 0; row < 20; row++) {
+        for (unsigned col = 0; col < 10; col++) {
             const auto& cell = well.matrix.at(row + 2).at(col);
             if (cell) {
                 cell->draw(draw_offset_x + col * Mino::texture_size_px,
@@ -71,13 +71,13 @@ void Render::drawContent(const Well& well, GraphicsContext& gcx, int draw_offset
         }
 
         // draw piece
-        for (unsigned row = 0; row < 4; row++) {
+        for (int row = 0; row < 4; row++) {
             if (well.active_piece_y + row < 1) // hide row 22
                 continue;
 
             if (well.active_piece_y + row < 2) { // partially draw row 21
                 draw_offset_y -= top_row_height;
-                for (unsigned col = 0; col < 4; col++) {
+                for (int col = 0; col < 4; col++) {
                     const auto& cell = well.active_piece->currentGrid().at(row).at(col);
                     if (cell) {
                         cell->drawPartial(top_row_cliprect, {
